ai-content-empire/requirements.txt
huggingface_hub
requests
PyYAML
psycopg2-binary
supabase
